requirements.txt
django>=5.1,<5.2
djangorestframework>=3.14,<3.15
django-cors-headers>=4.3,<4.4
requests>=2.31,<2.32
python-dotenv>=1.0,<1.1
django-filter>=23.5,<24.0
drf-yasg>=1.21.7,<1.22
pytest>=7.4.3,<7.5
pytest-django>=4.7.0,<4.8
gunicorn>=21.2.0,<21.3
whitenoise>=6.6.0,<6.7
dj-database-url>=2.1.0,<2.2
coreapi>=2.3.3,<2.4
PyJWT==2.8.0 